folders, careers page static, job cards 3 per row
